Fix Refatorando Controller de PurchaseController, e terminando de implementar o service de PurchaseService, verificando regras de negocio
