+++ src/main/java/br/com/meli/grupo1/desafio_spring/DTO/purchases/Purchase.java
@@ -1,0 +1,40 @@
+package br.com.meli.grupo1.desafio_spring.DTO.purchases;
+
+public class Purchase {
+    private Integer productId;
+    private String name;
+    private String brand;
+    private Integer quantity;
+
+    public Integer getProductId() {
+        return productId;
+    }
+
+    public void setProductId(Integer productId) {
+        this.productId = productId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getBrand() {
+        return brand;
+    }
+
+    public void setBrand(String brand) {
+        this.brand = brand;
+    }
+
+    public Integer getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(Integer quantity) {
+        this.quantity = quantity;
+    }
+}
