+++ src/main/java/br/com/meli/grupo1/desafio_spring/controllers/PurchaseController.java
@@ -1,13 +1,13 @@
 package br.com.meli.grupo1.desafio_spring.controllers;
 
-import br.com.meli.grupo1.desafio_spring.DTO.PurchaseRequestDTO;
+import br.com.meli.grupo1.desafio_spring.DTO.purchases.ListOrderDTO;
+import br.com.meli.grupo1.desafio_spring.DTO.purchases.PurchaseRequestDTO;
+import br.com.meli.grupo1.desafio_spring.services.PurchaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 public class PurchaseController {
@@ -16,8 +16,8 @@ public class PurchaseController {
     private PurchaseService purchaseService;
 
     @PostMapping("/purchase-request")
-    public ResponseEntity<Order> purchaseOrder(@RequestBody PurchaseRequestDTO products){
-        Order orderPurchase = purchaseService.purchaseOrder(products);
-        ResponseEntity.ok().body(orderPurchase);
+    public ResponseEntity<ListOrderDTO> purchaseOrder(@RequestBody PurchaseRequestDTO products){
+        ListOrderDTO orderPurchase = purchaseService.purchaseOrder(products);
+        return ResponseEntity.ok().body(orderPurchase);
     }
 }
