+++ src/main/java/br/com/meli/grupo1/desafio_spring/repositories/PurchaseRepository.java
@@ -1,9 +1,12 @@
 package br.com.meli.grupo1.desafio_spring.repositories;
 
-import br.com.meli.grupo1.desafio_spring.DTO.PurchaseRequestDTO;
+import br.com.meli.grupo1.desafio_spring.entities.Product;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 @Repository
 public interface PurchaseRepository {
     public boolean existProduct(Integer id);
+    public List<Product> findProductsIn(List<Integer> ids);
 }
